made tile throwable
To go to throwing part of game you must press spacebar, then when you click on the tile it will launch forward and upward at a set velocity. Need to find a way for the user to set the velocity but once that is found it will be very easy to implement to the throwing. The tile that is thrown is a completely separate tile and will be its own thing. I believe we can have it copy the texture of the tile that is meant to be thrown and it will look like they are throwing that tile. Going to find a way to reset tile after a certain time of it launching.